Overhaul scanning
Reset values before scanning only if it was
not set manually

Still scan in and filename if values were set manually.
Set values and in_idx only if it was not set manually.

Set matches according to already set values (manually or by scanning).
Values do not really need to be sorted for it to work.
Values must be unique though